Enter downloadmanagement.{h,cpp}
Moved DownloadInfo and DownloadState into new files
downloadmanagement.{h,cpp}.
